COMP: Fixed warnings.

+++ VTK/Utilities/vtkexodus2/exgblk.c
@@ -50,7 +50,7 @@
 *
 * revision history -
 *
-*  $Id: exgblk.c,v 1.2 2009-01-16 14:32:01 utkarsh Exp $
+*  $Id: exgblk.c,v 1.3 2009-01-29 15:13:41 berk Exp $
 *
 */
 
@@ -275,7 +275,7 @@ int ex_get_block( int exoid,
       len = MAX_STR_LENGTH;
       sprintf (errmsg,
 	       "Warning: %s %d type will be truncated to %ld chars", 
-	       ex_name_of_object(blk_type), blk_id,len);
+	       ex_name_of_object(blk_type), blk_id, (long)(len));
       ex_err("ex_get_block",errmsg,EX_MSG);
     }
